updates to structure, [id].tsx, moved data types into types
