+++ src/app/dashboard/dashboard-pie-chart/dashboard-pie-chart.tsx
@@ -3,7 +3,7 @@
 import { PieChart } from '@metrostar/comet-data-viz';
 import { Spacecraft } from '../../../types/spacecraft';
 import React, { useEffect, useState } from 'react';
-import { ChartData } from '../types';
+import { ChartData } from '../../../types/data';
 
 interface DashboardPieChartProps {
   items: Spacecraft[] | undefined;
