+++ src/app/dashboard/dashboard-table/dashboard-table.test.tsx
@@ -4,7 +4,7 @@ import { BrowserRouter } from 'react-router-dom';
 import { RecoilRoot } from 'recoil';
 import { DashboardTable } from './dashboard-table';
 import { vi } from 'vitest';
-import { Spacecraft } from 'spacecraft';
+//import { Spacecraft } from 'spacecraft';
 
 
 // Mock the DataTable component
