+++ src/app/details/details.test.tsx
@@ -2,7 +2,7 @@
 import React from 'react';
 import { render } from '@testing-library/react';
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
-import Details from './page';
+import Details from './[id]';
 import { vi } from 'vitest';
 
 // Mocking the hooks
